+++ lib/govuk_app_config/govuk_json_logging.rb
@@ -1,32 +1,31 @@
+require "json"
 require "logstasher"
 require "action_controller"
-require_relative "rails_ext/action_dispatch/debug_exceptions"
 
 module GovukJsonLogging
   def self.configure
-    # GOV.UK Rails applications are expected to output JSON to stdout which is
-    # then indexed in a Kibana instance. These log outputs are created by the
-    # logstasher gem.
-    #
-    # Rails applications will typically write other things to stdout such as
-    # `Rails.logger` calls or 'puts' statements. However these are not in a
-    # JSON format which causes problems for the log file parsers.
-    #
-    # To resolve this we redirect stdout to stderr, to cover any Rails
-    # writing. This frees up the normal stdout for the logstasher logs.
-    #
-    # We also disable buffering, so that logs aren't lost on crash or delayed
+    # We disable buffering, so that logs aren't lost on crash or delayed
     # indefinitely while troubleshooting.
-
-    # rubocop:disable Style/GlobalVars
-    $real_stdout = $stdout.clone
-    $real_stdout.sync = true
-    $stdout.reopen($stderr)
     $stdout.sync = true
-    # rubocop:enable Style/GlobalVars
 
-    # Send Rails' logs to STDERR because they're not JSON formatted.
-    Rails.logger = ActiveSupport::TaggedLogging.new(Logger.new($stderr, level: Rails.logger.level))
+    Rails.logger = Logger.new(
+      $stdout,
+      level: Rails.logger.level,
+      formatter: proc { |severity, datetime, _progname, msg|
+        hash = {
+          "@timestamp": datetime.utc.iso8601(3),
+          message: msg,
+          level: severity,
+          tags: %w[rails],
+        }
+
+        if defined?(GdsApi::GovukHeaders) && !GdsApi::GovukHeaders.headers[:govuk_request_id].nil?
+          hash[:govuk_request_id] = GdsApi::GovukHeaders.headers[:govuk_request_id]
+        end
+
+        "#{hash.to_json}\n"
+      },
+    )
 
     LogStasher.add_custom_fields do |fields|
       # Mirrors Nginx request logging, e.g. GET /path/here HTTP/1.1
@@ -53,7 +52,7 @@ module GovukJsonLogging
     Rails.application.config.logstasher.source = {}
 
     Rails.application.config.logstasher.logger = Logger.new(
-      $real_stdout, # rubocop:disable Style/GlobalVars
+      $stdout,
       level: Rails.logger.level,
       formatter: proc { |_severity, _datetime, _progname, msg|
         "#{msg.is_a?(String) ? msg : msg.inspect}\n"
@@ -68,7 +67,5 @@ module GovukJsonLogging
       # the responses it gets, so direct this to the logstasher logger.
       GdsApi::Base.default_options[:logger] = Rails.application.config.logstasher.logger
     end
-
-    RailsExt::ActionDispatch.monkey_patch_log_error if RailsExt::ActionDispatch.should_monkey_patch_log_error?
   end
 end